+++ src/main/java/com/example/jingangfarmmanagement/model/req/MealVoucherMaterialReq.java
@@ -9,5 +9,5 @@ import lombok.*;
 @Setter
 public class MealVoucherMaterialReq {
     Long materialId;
-    Long quantity;
+    Double quantity;
 }
